edit, list view update

+++ src/pages/List.vue
@@ -6,15 +6,16 @@
             <th>ID</th>
             <th>Ism</th>
             <th>Mutaxassisligi</th>
-            <th>Batafsil</th>
-            <th>O'chirish</th>
+            <th></th>
         </tr>
         <tr v-for='(doctor,index) of doctors' :key='index'>
             <td>{{doctor.id}}</td>
             <td>{{doctor.name}}</td>
             <td>{{doctor.spec}}</td>
-            <td><router-link :to="'/view/'+doctor.id">Batafsil</router-link></td>
-            <td><button class="del" @click='del(index)'>Delet</button></td>
+            <td>
+                <router-link :to="'/view/'+doctor.id" class="icon-btn"><i class="fas fa-eye"></i></router-link>
+                <router-link :to="'/edit/'+doctor.id" class="icon-btn"><i class="fas fa-edit"></i></router-link>
+            <button class="del" @click='del(index)'> <i class="fas fa-trash"></i> </button></td>
         </tr>
     </table>
     </div>
@@ -30,10 +31,13 @@ export default {
     },
     methods:{
         del(index){
-            axios.delete('http://localhost:3000/doctors/'+this.doctors[index].id).then(response=>{
-                console.log(response.data)
-                this.doctors.splice(index,1)
-            })
+            if(confirm('Qaroringiz aniqmi?')){
+
+                axios.delete('http://localhost:3000/doctors/'+this.doctors[index].id).then(response=>{
+                    console.log(response.data)
+                    this.doctors.splice(index,1)
+                })
+            }
         }
     },
     created(){
@@ -42,13 +46,24 @@ export default {
 }
 </script>
 
-<style>
-.del{
-    background: rgb(255, 0, 0);
-    border: 0;
-    padding: 12px;
-    border-radius: 10px;
-    text-transform: uppercase;
+
+<style scoped>
+button{
+    background: unset;
+    border:0;
     cursor: pointer;
+}
+button:hover{
+    color: crimson;
+}
+.icon-btn{
+    color: #000;
+    margin-right: 10px;
+}
+.icon-btn:hover{
+    color: darkorange;
+}
+table tr td:last-child{
+    width: 100px;
 }
 </style>
